refactor: externalize subjects data
