+++ index.html
@@ -19,6 +19,6 @@
   </div>
 </div>
 <div id="final-overlay" style="display:none"></div>
-<script src="js/main.js"></script>
+<script type="module" src="js/main.js"></script>
 </body>
 </html>
