+++ js/main.js
@@ -1,14 +1,4 @@
-const subjects = {
-  '语文': ['背诵《春晓》', '完成练习册第10页'],
-  '数学': ['做作业本P20-22', '复习二次函数'],
-  '英语': ['背单词Unit5', '完成语法练习'],
-  '物理': ['预习第3章', '完成实验报告'],
-  '化学': ['练习化学方程式', '完成习题2-4'],
-  '生物': ['复习细胞结构', '完成小测卷'],
-  '历史': ['阅读《中国近代史》第2章', '整理课堂笔记'],
-  '政治': ['思考题第1-3题'],
-  '地理': ['完成地图册练习']
-};
+import subjects from '../data/subjects.js';
 const preview = document.getElementById('preview');
 let totalTasks = 0;
 for (const [subject, works] of Object.entries(subjects)) {
